requirements.txt
asgiref==3.4.1
boto3==1.20.5
botocore==1.23.5
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
django-storages==1.12
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
Pillow==8.4.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
